third_party/blink/web_tests/fast/loader/stateobjects/pushstate-in-data-url-denied-expected.txt

ALERT: PASS: data URLs cannot be manipulated via pushState.
